pyproject.toml
[project]
name = "artifact-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langgraph>=1.0.6",
    "mcp>=1.25.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
]
